refactor 불필요한 변수선언 수정(MemberUpdater) && MemberCreator 불필요한 변수선언 제거
