+++ app/src/main/java/com/codesoom/assignment/application/member/MemberUpdater.java
@@ -1,7 +1,6 @@
 package com.codesoom.assignment.application.member;
 
 import com.codesoom.assignment.domain.Member;
-import com.codesoom.assignment.domain.MemberRepository;
 import com.codesoom.assignment.web.shop.member.dto.MemberUpdateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -9,8 +8,6 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class MemberUpdater {
-
-    private final MemberRepository memberRepository;
     private final MemberReader memberReader;
 
     public Member update(MemberUpdateRequest memberUpdateRequest) {
